% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_countdown.R
\name{format_time}
\alias{format_time}
\title{FUNCTION_TITLE}
\usage{
format_time(secs)
}
\arguments{
\item{secs}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
