% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_functions.R
\name{get_player_details}
\alias{get_player_details}
\title{FUNCTION_TITLE}
\usage{
get_player_details(bs_static)
}
\arguments{
\item{type}{PARAM_DESCRIPTION, Default: c("general", "gw")[1]}

\item{extra_cols}{PARAM_DESCRIPTION, Default: c("sample", "sample")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
