% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_functions.R
\name{get_url_data}
\alias{get_url_data}
\title{FUNCTION_TITLE}
\usage{
get_url_data(url)
}
\arguments{
\item{url}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
