% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{t_col}
\alias{t_col}
\title{FUNCTION_TITLE}
\usage{
t_col(colour, percent = 10, name = NULL)
}
\arguments{
\item{colour}{PARAM_DESCRIPTION}

\item{percent}{PARAM_DESCRIPTION, Default: 10}

\item{name}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
