% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_functions.R
\name{table_live_draft_picks}
\alias{table_live_draft_picks}
\title{FUNCTION_TITLE}
\usage{
table_live_draft_picks(league, plot_table = F)
}
\arguments{
\item{league}{PARAM_DESCRIPTION, Default: league_id}

\item{plot_table}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
